update tide.py and util.py
Successfully merged ```fetch_noaa_tide_data()```  into ```tide.py```
Improved the other functions to have similar format as ```fetch_noaa_tide_data()``` in ```tide.py```
Removed ```fetch_noaa_tide_data()``` from ```util.py``` but maintained a functional import statement for ```fetch_noaa_tide_data()```
